function moved to different file, added typing
